+++ stdlib/Distributed/src/cluster.jl
@@ -163,8 +163,9 @@ mutable struct LocalProcess
 end
 
 
+import LinearAlgebra
 function disable_threaded_libs()
-    BLAS.set_num_threads(1)
+    LinearAlgebra.BLAS.set_num_threads(1)
 end
 
 worker_timeout() = parse(Float64, get(ENV, "JULIA_WORKER_TIMEOUT", "60.0"))