+++ stdlib/SparseArrays/src/abstractsparse.jl
@@ -27,12 +27,12 @@ false
 issparse(A::AbstractArray) = false
 issparse(S::AbstractSparseArray) = true
 
-issparse(S::Symmetric{<:Any,<:AbstractSparseMatrix}) = true
-issparse(S::Hermitian{<:Any,<:AbstractSparseMatrix}) = true
-issparse(S::LowerTriangular{<:Any,<:AbstractSparseMatrix}) = true
-issparse(S::LinAlg.UnitLowerTriangular{<:Any,<:AbstractSparseMatrix}) = true
-issparse(S::UpperTriangular{<:Any,<:AbstractSparseMatrix}) = true
-issparse(S::LinAlg.UnitUpperTriangular{<:Any,<:AbstractSparseMatrix}) = true
+issparse(S::LinearAlgebra.Symmetric{<:Any,<:AbstractSparseMatrix}) = true
+issparse(S::LinearAlgebra.Hermitian{<:Any,<:AbstractSparseMatrix}) = true
+issparse(S::LinearAlgebra.LowerTriangular{<:Any,<:AbstractSparseMatrix}) = true
+issparse(S::LinearAlgebra.UnitLowerTriangular{<:Any,<:AbstractSparseMatrix}) = true
+issparse(S::LinearAlgebra.UpperTriangular{<:Any,<:AbstractSparseMatrix}) = true
+issparse(S::LinearAlgebra.UnitUpperTriangular{<:Any,<:AbstractSparseMatrix}) = true
 
 indtype(S::AbstractSparseArray{<:Any,Ti}) where {Ti} = Ti