+++ test/choosetests.jl
@@ -33,7 +33,7 @@ in the `choices` argument:
 """ ->
 function choosetests(choices = [])
     testnames = [
-        "linalg", "subarray", "core", "compiler", "worlds",
+        "subarray", "core", "compiler", "worlds",
         "keywordargs", "numbers", "subtype",
         "char", "strings", "triplequote", "unicode", "intrinsics",
         "dict", "hashing", "iobuffer", "staged", "offsetarray",
@@ -109,23 +109,6 @@ function choosetests(choices = [])
     elseif "subarray" in tests
         filter!(x -> x != "subarray", tests)
         prepend!(tests, ["subarray"])
-    end
-
-    linalgtests = ["linalg/triangular", "linalg/qr", "linalg/dense",
-                   "linalg/matmul", "linalg/schur", "linalg/special",
-                   "linalg/eigen", "linalg/bunchkaufman", "linalg/svd",
-                   "linalg/lapack", "linalg/tridiag", "linalg/bidiag",
-                   "linalg/diagonal", "linalg/pinv", "linalg/givens",
-                   "linalg/cholesky", "linalg/lu", "linalg/symmetric",
-                   "linalg/generic", "linalg/uniformscaling", "linalg/lq",
-                   "linalg/hessenberg", "linalg/blas", "linalg/adjtrans"]
-
-    if "linalg" in skip_tests
-        filter!(x -> (x != "linalg" && !(x in linalgtests)), tests)
-    elseif "linalg" in tests
-        # specifically selected case
-        filter!(x -> x != "linalg", tests)
-        prepend!(tests, linalgtests)
     end
 
     compilertests = ["compiler/compiler", "compiler/validation"]