+++ test/replutil.jl
@@ -1,6 +1,6 @@
 # This file is a part of Julia. License is MIT: https://julialang.org/license
 
-using Random
+using Random, LinearAlgebra
 
 # For curmod_*
 include("testenv.jl")