+++ NEWS.md
@@ -899,6 +899,9 @@ Deprecated or removed
 
   * Sparse array functionality has moved to the `SparseArrays` standard library module ([#25249]).
 
+  * Linear algebra functionality, and specifically the `LinAlg` module has moved to the
+    `LinearAlgebra` standard library module ([#25571]).
+
   * `@printf` and `@sprintf` have been moved to the `Printf` standard library ([#23929],[#25056]).
 
   * The aliases `Complex32`, `Complex64` and `Complex128` have been deprecated in favor of `ComplexF16`,